Refactor quote handling
* Shift quote handling code out from languages.rs to quotes.rs
* Apply cache and scan optimisations to quote searching.
* Fix - a closer hugging a terminator stays an orphan.
* Fix - `quote_candidates_should_pair` keeps pairing back to back spans unless the closer cleanly opens a further quote ahead.
