+++ src/languages/language.rs
@@ -7,12 +7,14 @@ use crate::constants::EMAIL_REGEX;
 use crate::constants::EXCLAMATION_WORDS;
 use crate::constants::GLOBAL_SENTENCE_TERMINATORS;
 use crate::constants::PARENS_REGEX;
-use crate::constants::QUOTE_CLOSERS_BY_LEN;
-use crate::constants::QUOTE_PAIRS;
-use crate::constants::QUOTES_REGEX;
 use crate::constants::QuotePair;
-use crate::constants::SPACE_AFTER_SEPARATOR;
 use crate::constants::is_sentence_terminator;
+
+use super::quotes::{
+    OrphanCloserPositions, QuoteMispairing, collect_quote_ranges, extend_past_orphan_closer,
+    inner_terminator_boundary, is_symmetric_quote_closer, is_symmetric_quote_mispairing,
+    peel_leading_symmetric_quote, tag_quote_mispairing,
+};
 
 use super::trailing_markers::{MarkerTable, classify_trailing_marker, marker_bypasses_suppression};
 
@@ -89,25 +91,6 @@ pub(crate) fn paragraph_breaks(text: &str) -> impl Iterator<Item = (usize, usize
     })
 }
 
-/// True iff it finds a `.[ \t]+[A-Z]` shape.
-/// Using memchr here instead of regex for speed
-fn has_possible_inline_sentence_break(span: &str) -> bool {
-    let bytes = span.as_bytes();
-    for dot_pos in memchr::memchr_iter(b'.', bytes) {
-        let tail = &bytes[dot_pos + 1..];
-
-        let Some(blanks) = tail.iter().position(|&b| !matches!(b, b' ' | b'\t')) else {
-            continue;
-        };
-
-        if blanks > 0 && tail[blanks].is_ascii_uppercase() {
-            return true;
-        }
-    }
-
-    false
-}
-
 /// True when a `.` sits inside a code-like numbered token rather than ending a
 /// sentence. A digit immediately before, and an alphanumeric token with a digit
 /// after with no space, e.g. the chess move `7.Bg5`.
@@ -155,374 +138,7 @@ fn starts_with_initial(s: &str) -> bool {
         && chars.next().is_none_or(char::is_whitespace)
 }
 
-/// True when `closer` is the closing token of a symmetric quote pair —
-/// `'`, `''`, `"`, etc., where opener equals closer. These pairs are
-/// inherently ambiguous: `QUOTES_REGEX` can fail to pair them when they're
-/// space-padded, so callers need extra logic to decide orphanhood.
-fn is_symmetric_quote_closer(closer: &str) -> bool {
-    QUOTE_PAIRS
-        .iter()
-        .any(|p| p.open == p.close && p.close == closer)
-}
-
-/// True when the `closer` token at `boundary` in `paragraph` is an orphan —
-/// either it has no matching opener earlier in the paragraph, or it is a
-/// lone stray with no counterpart anywhere.
-///
-/// Asymmetric closers (`»`, `”`, …) are unambiguous and always orphan once
-/// `QUOTES_REGEX` has had a chance to pair them. For symmetric closers we
-/// inspect the closer's distribution across the paragraph, ignoring
-/// occurrences already consumed by a paired quote range and contractions
-/// like `wasn't`.
-fn is_orphan_closer(
-    paragraph: &str,
-    boundary: usize,
-    closer: &str,
-    skippable_ranges: &[SkippableRange],
-) -> bool {
-    if !is_symmetric_quote_closer(closer) {
-        return true;
-    }
-
-    let (mut before, mut at_or_after) = (0usize, 0usize);
-    for (idx, _) in paragraph.match_indices(closer) {
-        if !is_quote_candidate(paragraph, idx, closer, skippable_ranges) {
-            continue;
-        }
-
-        if idx < boundary {
-            before += 1;
-        } else {
-            at_or_after += 1;
-        }
-    }
-
-    let unmatched_opener_before = before % 2 == 1;
-    let lone_stray_at_boundary = before == 0 && at_or_after == 1;
-    unmatched_opener_before || lone_stray_at_boundary
-}
-
-/// Find the first `QUOTE_PAIRS` entry whose `open` is a prefix of `span`.
-fn identify_quote_pair(span: &str) -> Option<&'static QuotePair> {
-    QUOTE_PAIRS.iter().find(|p| span.starts_with(p.open))
-}
-
-/// Single-char quote openers whose `open`/`close` contain no ASCII `'` or backtick.
-/// Map to `QuotePair`.
-static CLEAN_QUOTE_OPENERS: LazyLock<Vec<(char, &'static QuotePair)>> = LazyLock::new(|| {
-    let is_clean = |s: &str| !s.contains(['\'', '`']);
-
-    QUOTE_PAIRS
-        .iter()
-        .filter(|p| is_clean(p.open) && is_clean(p.close))
-        .filter_map(|p| {
-            let mut chars = p.open.chars();
-            let c = chars.next()?;
-            chars.next().is_none().then_some((c, p))
-        })
-        .collect()
-});
-
-/// The clean pair opened by `c`
-fn clean_opener_pair(c: char) -> Option<&'static QuotePair> {
-    CLEAN_QUOTE_OPENERS
-        .iter()
-        .find_map(|&(opener, pair)| (opener == c).then_some(pair))
-}
-
-/// Fast path when the text has no `'` or backtick and no `0xE3` lead byte (CJK variants).
-fn scan_unambiguous_quotes(text: &str, out: &mut Vec<SkippableRange>) {
-    let bytes = text.as_bytes();
-    let mut cursor = 0;
-
-    while let Some(rel) = memchr::memchr3(b'"', 0xC2, 0xE2, &bytes[cursor..]) {
-        let opener = cursor + rel;
-
-        let c = text[opener..]
-            .chars()
-            .next()
-            .expect("a lead byte cannot be at end of text");
-
-        let Some(pair) = clean_opener_pair(c) else {
-            cursor = opener + c.len_utf8();
-            continue;
-        };
-
-        let content_start = opener + pair.open.len();
-        match text[content_start..].find(pair.close) {
-            Some(off) => {
-                let end = content_start + off + pair.close.len();
-                out.push(SkippableRange::new_quote(opener, end, pair));
-                cursor = end;
-            }
-
-            // Opener with no closer
-            None => cursor = content_start,
-        }
-    }
-}
-
-/// True when `range` was constructed for a symmetric-pair quote token,
-/// e.g., `''…''`, `'…'`, `"…"`.
-fn is_symmetric_quote_range(range: &SkippableRange) -> bool {
-    range.quote_pair.is_some_and(|p| p.open == p.close)
-}
-
-/// Last symmetric quote token's whole paragraph occurrence parity.
-#[derive(Default)]
-struct ParityCache {
-    last: Option<(&'static str, bool)>,
-}
-
-impl ParityCache {
-    fn token_count_is_odd(&mut self, paragraph: &str, token: &'static str) -> bool {
-        if let Some((seen, odd)) = self.last
-            && seen == token
-        {
-            return odd;
-        }
-
-        let odd = paragraph.matches(token).count() % 2 == 1;
-        self.last = Some((token, odd));
-
-        odd
-    }
-}
-
-/// True when the paragraph contains an odd number of the symmetric quote
-/// token that opens `range`. An odd count guarantees at least one orphan
-/// occurrence — and when that orphan sits earlier than a real downstream
-/// opener, `QUOTES_REGEX` will mispair across a real sentence break. Even
-/// counts are structurally consistent and should be trusted.
-fn symmetric_token_count_is_odd(
-    paragraph: &str,
-    range: &SkippableRange,
-    cache: &mut ParityCache,
-) -> bool {
-    let Some(pair) = range.quote_pair else {
-        return false;
-    };
-
-    cache.token_count_is_odd(paragraph, pair.open)
-}
-
-/// The `parens` are sorted, mutually non-overlapping.
-/// Find the one containing byte offset `x` (`p.start < x < p.end`).
-fn paren_containing(parens: &[SkippableRange], x: usize) -> Option<&SkippableRange> {
-    parens
-        .partition_point(|p| p.start < x)
-        .checked_sub(1)
-        .map(|i| &parens[i])
-        .filter(|p| x < p.end)
-}
-
-/// True when `quote` partially overlaps any paren in `parens`, i.e., one part in, one part out.
-/// Full containment returns false.
-/// `parens` must be sorted by start and disjoint.
-fn quote_partially_overlaps_parens(quote: &SkippableRange, parens: &[SkippableRange]) -> bool {
-    debug_assert!(
-        parens.windows(2).all(|w| w[0].end <= w[1].start),
-        "parens must be sorted and non-overlapping"
-    );
-
-    if let Some(p) = paren_containing(parens, quote.start)
-        && p.end < quote.end
-    {
-        return true;
-    }
-
-    if let Some(p) = paren_containing(parens, quote.end)
-        && p.start > quote.start
-    {
-        return true;
-    }
-
-    false
-}
-
-/// Trim leading whitespace from `s`, then if a symmetric quote closer
-/// follows, peel one such closer plus its trailing whitespace. Used to
-/// look "through" a stray closing apostrophe when checking for a comma
-/// continuation, e.g., `. ' , Tim …`.
-fn peel_leading_symmetric_quote(s: &str) -> &str {
-    let trimmed = s.trim_start();
-    let Some(&first) = trimmed.as_bytes().first() else {
-        return trimmed;
-    };
-
-    // Fast check for ASCII quotes
-    if first.is_ascii() && !matches!(first, b'\'' | b'"' | b'`') {
-        return trimmed;
-    }
-
-    QUOTE_PAIRS
-        .iter()
-        .filter(|p| p.open == p.close)
-        .find_map(|p| trimmed.strip_prefix(p.close))
-        .map(str::trim_start)
-        .unwrap_or(trimmed)
-}
-
-/// True when `idx` lies inside an existing quote `SkippableRange`.
-/// Ranges are sorted so use a binary search to filter out irrelevant ranges.
-fn is_in_quote_range(ranges: &[SkippableRange], idx: usize) -> bool {
-    let pos = ranges.partition_point(|r| r.start <= idx);
-    ranges[..pos].iter().any(|r| r.is_quote() && idx < r.end)
-}
-
-/// Chars immediately before `idx` and immediately after `idx + token.len()`.
-/// Returned as `(prev, next)`; either side is `None` at the text boundary.
-fn neighbors(text: &str, idx: usize, token: &str) -> (Option<char>, Option<char>) {
-    (
-        text[..idx].chars().next_back(),
-        text[idx + token.len()..].chars().next(),
-    )
-}
-
-/// True when the `token` occurrence at `idx` is a contraction (`wasn't`,
-/// `o'clock`) sandwiched between two alphanumerics. Such `'` are not quote
-/// characters. Counting them flips the parity-based orphan check.
-fn is_contraction_quote(text: &str, idx: usize, token: &str) -> bool {
-    let (prev, next) = neighbors(text, idx, token);
-    prev.is_some_and(|c| c.is_alphanumeric()) && next.is_some_and(|c| c.is_alphanumeric())
-}
-
-/// True when the `token` occurrence at `idx` is a free-standing quote
-/// candidate not already inside a paired quote range and not a contraction.
-fn is_quote_candidate(text: &str, idx: usize, token: &str, ranges: &[SkippableRange]) -> bool {
-    !is_in_quote_range(ranges, idx) && !is_contraction_quote(text, idx, token)
-}
-
-/// True when the `'` or `` ` `` at `idx` is preceded by start of text or
-/// whitespace AND followed by whitespace. The shape `QUOTES_REGEX`'s guarded
-/// pattern can't pair. It requires `\b` after the opener.
-fn is_opener_shape(text: &str, idx: usize, token: &str) -> bool {
-    let (prev, next) = neighbors(text, idx, token);
-    prev.is_none_or(char::is_whitespace) && next.is_some_and(char::is_whitespace)
-}
-
-/// True when `text[from..]` starts with one or more ASCII blanks followed by
-/// an ASCII uppercase letter, the signature of a fresh utterance opening.
-fn starts_new_utterance(text: &str, from: usize) -> bool {
-    let tail = &text[from..];
-    let trimmed = tail.trim_start_matches([' ', '\t']);
-
-    trimmed.len() < tail.len()
-        && trimmed
-            .as_bytes()
-            .first()
-            .is_some_and(u8::is_ascii_uppercase)
-}
-
-/// True when the candidate `'` or `` ` `` at `opener` should pair with the
-/// candidate at `closer` into a quote range. Rejects two shapes:
-/// * `opener` isn't space-padded. It's not opener-like to begin with.
-/// * The span looks like two back to back utterances rather than one
-///   multi-sentence quotation. The closer itself starts a fresh utterance
-///   (whitespace + uppercase) AND the intervening text contains an inline
-///   `[.!] + ws + uppercase` sentence break.
-fn quote_candidates_should_pair(text: &str, opener: usize, closer: usize, token: &str) -> bool {
-    if !is_opener_shape(text, opener, token) {
-        return false;
-    }
-
-    let span = &text[opener + token.len()..closer];
-    let after_closer = closer + token.len();
-    let back_to_back =
-        starts_new_utterance(text, after_closer) && has_possible_inline_sentence_break(span);
-
-    !back_to_back
-}
-
-/// Define quote mispairings to improve readability in the orphan quote handling.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum QuoteMispairing {
-    #[default]
-    None,
-    Certain,
-    Possible,
-}
-
-/// Tag each symmetric-pair quote range with a mispairing label.
-fn populate_quote_mispairing(paragraph: &str, ranges: &mut [SkippableRange]) {
-    let mut cache = ParityCache::default();
-
-    let parens: Vec<SkippableRange> = ranges
-        .iter()
-        .filter(|r| r.range_type == SkippableRangeType::Parentheses)
-        .copied()
-        .collect();
-
-    for slot in ranges.iter_mut() {
-        if !slot.is_quote() {
-            continue;
-        }
-
-        let range = *slot;
-
-        let class = if !is_symmetric_quote_range(&range) {
-            QuoteMispairing::None
-        } else if quote_partially_overlaps_parens(&range, &parens) {
-            QuoteMispairing::Certain
-        } else if symmetric_token_count_is_odd(paragraph, &range, &mut cache) {
-            QuoteMispairing::Possible
-        } else {
-            QuoteMispairing::None
-        };
-
-        slot.quote_mispairing = class;
-    }
-}
-
-/// Append `'…'` / `` `…` `` ranges that `QUOTES_REGEX` couldn't pair.
-/// Guarded patterns require `\b` immediately after the opener, so
-/// space padded openers like `' word ` go unpaired even when they form a
-/// real `' … '` pair.
-fn append_space_padded_quote_pairs(text: &str, ranges: &mut Vec<SkippableRange>) {
-    let bytes = text.as_bytes();
-
-    // Fast check
-    if memchr::memchr2(b'\'', b'`', bytes).is_none() {
-        return;
-    }
-
-    for pair in QUOTE_PAIRS
-        .iter()
-        .filter(|p| p.ambiguous && p.open == p.close)
-    {
-        let token = pair.close;
-        debug_assert_eq!(
-            token.len(),
-            1,
-            "ambiguous symmetric tokens are single-byte ASCII"
-        );
-
-        if !bytes.contains(&token.as_bytes()[0]) {
-            continue;
-        }
-
-        let mut pending: Option<usize> = None;
-        for (idx, _) in text.match_indices(token) {
-            if !is_quote_candidate(text, idx, token, ranges) {
-                continue;
-            }
-
-            if let Some(opener) = pending
-                && quote_candidates_should_pair(text, opener, idx, token)
-            {
-                let end = idx + token.len();
-                ranges.push(SkippableRange::new_quote(opener, end, pair));
-                pending = None;
-            } else {
-                pending = Some(idx);
-            }
-        }
-    }
-}
-
 /// Push `boundary` only if it advances past the last recorded position.
-/// Quote-extension can move a boundary past later regex matches in the same
-/// paragraph; this keeps the boundary list strictly increasing.
 /// Assumes the caller is passing a non empty list of boundaries.
 fn push_if_increasing(boundaries: &mut Vec<usize>, boundary: usize) {
     debug_assert!(!boundaries.is_empty());
@@ -661,6 +277,7 @@ struct ParagraphScratch {
     sentence_boundaries: Vec<usize>,
     matches: Vec<(usize, usize)>,
     skippable_ranges: Vec<SkippableRange>,
+    orphan_closers: OrphanCloserPositions,
 }
 
 impl ParagraphScratch {
@@ -669,6 +286,7 @@ impl ParagraphScratch {
             sentence_boundaries: Vec::with_capacity(capacity),
             matches: Vec::with_capacity(capacity),
             skippable_ranges: Vec::with_capacity(capacity),
+            orphan_closers: OrphanCloserPositions::default(),
         }
     }
 }
@@ -740,9 +358,9 @@ fn push_paragraph_sentences<'a>(
 /// `binary_search` is whether to use binary search when searching in the ranges, i.e., is there enough data to
 /// justify the overhead of binary search.
 #[derive(Clone, Copy)]
-struct NonListRegion {
-    len: usize,
-    binary_search: bool,
+pub(crate) struct NonListRegion {
+    pub(crate) len: usize,
+    pub(crate) binary_search: bool,
 }
 
 /// Minimum number of ranges required before it's worth while to use binary search (rather than a linear scan).
@@ -759,7 +377,10 @@ fn collect_sentence_breaks<L: Language + ?Sized>(
         sentence_boundaries,
         matches,
         skippable_ranges,
+        orphan_closers,
     } = scratch;
+
+    orphan_closers.reset();
 
     sentence_boundaries.clear();
     sentence_boundaries.push(0);
@@ -791,7 +412,14 @@ fn collect_sentence_breaks<L: Language + ?Sized>(
             non_list_region,
         ) {
             Some(range) => inner_terminator_boundary(lang, paragraph, range, boundary),
-            None => Some(lang.extend_past_orphan_closer(paragraph, boundary, skippable_ranges)),
+            None => Some(extend_past_orphan_closer(
+                lang,
+                paragraph,
+                boundary,
+                skippable_ranges,
+                non_list_region,
+                orphan_closers,
+            )),
         };
 
         if let Some(break_at) = break_at {
@@ -854,7 +482,7 @@ fn containing_range<'r, L: Language + ?Sized>(
 ) -> Option<&'r SkippableRange> {
     let is_break = |range: &SkippableRange| {
         range.contains(boundary)
-            && !lang.is_symmetric_quote_mispairing(paragraph, range, match_start, match_end)
+            && !is_symmetric_quote_mispairing(lang, paragraph, range, match_start, match_end)
     };
 
     if region.binary_search {
@@ -862,23 +490,6 @@ fn containing_range<'r, L: Language + ?Sized>(
     } else {
         ranges.iter().find(|&r| is_break(r))
     }
-}
-
-/// The offset just past `range`'s closer where the sentence resumes, when `boundary`
-/// is a terminator at the range's inner edge (e.g. the . in "... end."), otherwise return `None`.
-fn inner_terminator_boundary<L: Language + ?Sized>(
-    lang: &L,
-    paragraph: &str,
-    range: &SkippableRange,
-    boundary: usize,
-) -> Option<usize> {
-    if !range.is_inner_terminator(paragraph, boundary) {
-        return None;
-    }
-
-    let next_word = lang.get_next_word_approx(paragraph, range.end);
-    lang.get_boundary_extend(next_word)
-        .map(|extend| range.end + extend)
 }
 
 /// Push a skippable range for each list-item line span (the last to `paragraph_len`)
@@ -1001,25 +612,6 @@ impl SkippableRange {
 
     pub fn is_quote(&self) -> bool {
         self.range_type == SkippableRangeType::Quote
-    }
-
-    /// True if `boundary` lies just before this quote's closing mark — i.e. the
-    /// sentence terminator sits inside the quoted span with only the closer left.
-    /// Opinionated behaviour, but it can help to resolve some real world cases with erroneous
-    /// or ambiguous punctuation placement.
-    ///
-    /// Example: in (start of line) `He said "Hello."`, the `.` boundary is immediately followed
-    /// by the closing `"`, so the sentence should extend past the quote rather than break
-    /// between the `.` and the `"`.
-    pub fn is_inner_terminator(&self, text: &str, boundary: usize) -> bool {
-        if !self.is_quote() || boundary >= self.end {
-            return false;
-        }
-
-        let head = &text[..self.end];
-        QUOTE_CLOSERS_BY_LEN
-            .iter()
-            .any(|c| head.ends_with(*c) && boundary + c.len() == self.end)
     }
 }
 
@@ -1184,80 +776,6 @@ pub trait Language {
         }
 
         Some(count)
-    }
-
-    /// If `boundary` sits at an orphan trailing quote closer (e.g. `.'` with no
-    /// matching opener captured by `QUOTES_REGEX`), advance past the closer, any
-    /// trailing whitespace, and any stranded terminator that would otherwise
-    /// form a single-punctuation sentence. Returns `boundary` unchanged otherwise.
-    fn extend_past_orphan_closer(
-        &self,
-        paragraph: &str,
-        boundary: usize,
-        skippable_ranges: &[SkippableRange],
-    ) -> usize {
-        // If the next char opens a known quoted range, that quote belongs to
-        // the upcoming sentence — leave the boundary alone.
-        if skippable_ranges.iter().any(|r| r.start == boundary) {
-            return boundary;
-        }
-
-        // Find an orphan closer starting at `boundary`, if any. Longest first
-        // so `''` wins over `'` when both could match.
-        let Some(closer) = QUOTE_CLOSERS_BY_LEN.iter().find(|c| {
-            paragraph[boundary..].starts_with(**c)
-                && is_orphan_closer(paragraph, boundary, c, skippable_ranges)
-        }) else {
-            return boundary;
-        };
-
-        // A symmetric closer (`''`, `'`, `"`, …) that follows a terminator+space
-        // and precedes whitespace + a capitalized word is more plausibly the
-        // *opener* of the next sentence than a trailing orphan — but only when
-        // the same token has already been used as a paired opener/closer
-        // earlier in the paragraph. That paired use is the signal that the
-        // text is using this token in opener position too; without it, the
-        // token is more likely a stray closer (e.g. `… do ? ''` with no
-        // earlier opener).
-        if is_symmetric_quote_closer(closer) {
-            let has_earlier_symmetric_pair = skippable_ranges.iter().any(|r| {
-                r.end <= boundary
-                    && is_symmetric_quote_range(r)
-                    && paragraph[r.start..].starts_with(*closer)
-            });
-
-            if has_earlier_symmetric_pair {
-                let after = &paragraph[boundary + closer.len()..];
-                let mut chars = after.chars();
-                let first = chars.next();
-                if first.is_some_and(char::is_whitespace) {
-                    let next_non_ws = chars.find(|c| !c.is_whitespace());
-                    if next_non_ws.is_some_and(|c| c.is_ascii_uppercase()) {
-                        return boundary;
-                    }
-                }
-            }
-        }
-
-        let advance_past_space = |pos: usize| {
-            SPACE_AFTER_SEPARATOR
-                .find(&paragraph[pos..])
-                .map_or(pos, |m| pos + m.end())
-        };
-
-        let mut boundary = advance_past_space(boundary + closer.len());
-        let sentence_break_regex = self.get_sentence_break_regex();
-
-        // Absorb any stranded terminators (e.g. `'' .`) that would otherwise
-        // form a single-punctuation sentence.
-        while let Some(m) = sentence_break_regex
-            .find(&paragraph[boundary..])
-            .filter(|m| m.start() == 0)
-        {
-            boundary = advance_past_space(boundary + m.end());
-        }
-
-        boundary
     }
 
     /// Checks if a potential sentence boundary is actually part of an abbreviation.
@@ -1437,28 +955,6 @@ pub trait Language {
             .any(|w| w.strip_suffix('!').is_some_and(|p| p == last_word))
     }
 
-    /// True when this symmetric-pair quote range (`''…''`, `'…'`, `"…"`) is
-    /// a likely `QUOTES_REGEX` mispairing across a sentence break.
-    /// - Certain: range straddles a parens boundary. Override always
-    ///   fires. A quote pair shouldn't cross a parens edge.
-    /// - Possible: paragraph has an odd token count (one or more
-    ///   orphans). Override fires only if `[start, end)` also looks like
-    ///   a strong sentence break.
-    /// - None: not a symmetric pair or evenly balanced. No override.
-    fn is_symmetric_quote_mispairing(
-        &self,
-        paragraph: &str,
-        range: &SkippableRange,
-        start: usize,
-        end: usize,
-    ) -> bool {
-        match range.quote_mispairing {
-            QuoteMispairing::None => false,
-            QuoteMispairing::Certain => true,
-            QuoteMispairing::Possible => self.has_strong_sentence_break(paragraph, start, end),
-        }
-    }
-
     /// True when the terminator at `[start, end)` looks like a confident
     /// sentence end: a single `.` whose preceding word is a symmetric quote
     /// closer (`''`, `"`, …) and whose follower starts with a capital letter
@@ -1630,21 +1126,7 @@ pub trait Language {
     fn get_skippable_ranges(&self, text: &str, out: &mut Vec<SkippableRange>) {
         out.clear();
 
-        // Fast path: ASCII `'`, backtick, and the `0xE3` lead byte are the only
-        // chars in the ambiguous and CJK (`《》「」`) quote pairs.
-        // If those are not present, only `open(?s:.*?)close` pairs can match.
-        if memchr::memchr3(b'\'', b'`', 0xE3, text.as_bytes()).is_none() {
-            scan_unambiguous_quotes(text, out);
-        } else {
-            for mat in QUOTES_REGEX.find_iter(text) {
-                let pair = identify_quote_pair(&text[mat.start()..])
-                    .expect("QUOTES_REGEX match must start with a known pair opener");
-
-                out.push(SkippableRange::new_quote(mat.start(), mat.end(), pair));
-            }
-
-            append_space_padded_quote_pairs(text, out);
-        }
+        collect_quote_ranges(text, out);
 
         for mat in PARENS_REGEX.find_iter(text) {
             out.push(SkippableRange::new(
@@ -1666,7 +1148,7 @@ pub trait Language {
         out.sort_unstable_by_key(|r| r.start);
 
         // Cache mispairing on each quote range for re-use
-        populate_quote_mispairing(text, out);
+        tag_quote_mispairing(text, out);
     }
 }
