+++ tests/en.txt
@@ -1132,3 +1132,46 @@ Give me 6 fl. oz. of milk.
 ---
 Give me 6 fl. oz. of milk.
 ===
+The point is,' said Tom, that I may not take that job.  There's other work coming up here on the Coast.' He paused, steadily eyeing Jed.
+---
+The point is,' said Tom, that I may not take that job.
+There's other work coming up here on the Coast.'
+He paused, steadily eyeing Jed. 
+===
+Jed read on to the conclusion: ' I unhesitatingly recommend him to be the best submarine inspector known to me for deep pier foundations.' 'Gosh, Tom, that's some recommendation!' Jed folded the letter and handed it back. ' But not a bit more than was coming to you,' he added. 'And no one knew that better than good old Bill Carey!' The point is,' said Tom, that I may not take that job. There's other work coming up here on the Coast.' He paused, steadily eyeing Jed. I was thinking,' he went on, 'that you are the one to take this foreign job. Fact is, Kid, I want you to have it.' For some minutes words wouldn't come to Jed. He was ready for any job - he knew that. But what was working by himself, even on a foreign job, compared to working with Tom King? Tom's own gospel, as he'd so often uttered it, answered him now - 'A diver must always aim to go deeper and deeper... he must always be ready to go when the call comes.'
+---
+Jed read on to the conclusion: ' I unhesitatingly recommend him to be the best submarine inspector known to me for deep pier foundations.'
+'Gosh, Tom, that's some recommendation!'
+Jed folded the letter and handed it back.
+' But not a bit more than was coming to you,' he added.
+'And no one knew that better than good old Bill Carey!'
+The point is,' said Tom, that I may not take that job.
+There's other work coming up here on the Coast.'
+He paused, steadily eyeing Jed.
+I was thinking,' he went on, 'that you are the one to take this foreign job. Fact is, Kid, I want you to have it.'
+For some minutes words wouldn't come to Jed.
+He was ready for any job - he knew that.
+But what was working by himself, even on a foreign job, compared to working with Tom King?
+Tom's own gospel, as he'd so often uttered it, answered him now - 'A diver must always aim to go deeper and deeper... he must always be ready to go when the call comes.'
+===
+You've got here just in time, Ivar,' he exclaimed.  `We can drive over the Bridge and take it easy, now. But in another week, traffic will be crowding over it. She's finished, all but a little more painting.'  ' It's a great yob, Tom!  The lad tried to tell me how it would look.  But seeing is believing! ' Tom nodded.  I guess he told you all about our diving job, too.  Eh, Jed?'
+---
+You've got here just in time, Ivar,' he exclaimed.
+`We can drive over the Bridge and take it easy, now. But in another week, traffic will be crowding over it. She's finished, all but a little more painting.'
+' It's a great yob, Tom!  The lad tried to tell me how it would look.  But seeing is believing! '
+Tom nodded.
+I guess he told you all about our diving job, too.
+Eh, Jed?'
+===
+Bill Carey, biting the end of his dead cigar, spoke quietly.  Red Dolan has gone over to get the Chief. ' Jed noticed, then, almost mechanically, that the Swan was heading back across the bay to her berth.
+---
+Bill Carey, biting the end of his dead cigar, spoke quietly.
+Red Dolan has gone over to get the Chief. '
+Jed noticed, then, almost mechanically, that the Swan was heading back across the bay to her berth.
+===
+Bill Carey, biting the end of his dead cigar, spoke quietly.  Red Dolan has gone over to get the Chief.' Jed noticed, then, almost mechanically, that the Swan was heading back across the bay to her berth.
+---
+Bill Carey, biting the end of his dead cigar, spoke quietly.
+Red Dolan has gone over to get the Chief.'
+Jed noticed, then, almost mechanically, that the Swan was heading back across the bay to her berth.
+===
